almost working node modules
