+++ src/resolver.js
@@ -1,0 +1,21 @@
+// can resolve the following paths and set to an absolute path
+// absolute
+function resolveAbsolute() {
+
+}
+// relative
+function resolveRelative() {
+
+}
+// node_modules
+function resolveModule() {
+    
+}
+
+function resolve() {
+    // if starts with  ./ or ../ then relative
+
+    // if starts with / or C:/ then absolute
+
+    // if anything else look in node modules
+}
